implementado abstractFacadeInterface

+++ SistemaTPI135-ejb/src/main/java/ues/edu/sv/mantenimientoTPI/acceso/AbstractFacade.java
@@ -11,9 +11,9 @@ import javax.persistence.EntityManager;
 /**
  *
  * @author esperanza
+ * @param <T>
  */
 public abstract class AbstractFacade<T> {
-
     private Class<T> entityClass;
 
     public AbstractFacade(Class<T> entityClass) {
@@ -22,17 +22,55 @@ public abstract class AbstractFacade<T> {
 
     protected abstract EntityManager getEntityManager();
 
-    public void create(T entity) {
-        getEntityManager().persist(entity);
+    public T create(T entity) {
+        if (entity != null && getEntityManager()!=null) {
+            getEntityManager().persist(entity);
+            return entity;
+        }else{
+            return null;
+        }
     }
 
-    public void edit(T entity) {
-        getEntityManager().merge(entity);
+    public T edit(T entity) {
+        if (entity != null && getEntityManager()!=null) {
+            getEntityManager().merge(entity);
+            return entity;
+        }else{
+            return null;
+        }
     }
 
-    public void remove(T entity) {
-        getEntityManager().remove(getEntityManager().merge(entity));
+    public T remove(T entity) {
+        if (entity!=null && getEntityManager()!=null) {
+            getEntityManager().remove(getEntityManager().merge(entity));
+            return entity;
+        }else{
+            return null;
+        }
     }
+    
+    public boolean crear(T entity){
+        
+        if (entity!= null && getEntityManager()!=null) {
+            return create(entity)==entity;
+        }
+        return false;
+    }
+    
+    public boolean modificar(T entity){
+        if (entity!= null && getEntityManager()!=null) {
+            return edit(entity)==entity;
+        }
+        return false;
+    }
+    
+    public boolean eliminar(T entity){
+        if (entity!= null && getEntityManager()!=null) {
+            return remove(entity)==entity;
+        }
+        return false;
+    }
+    
 
     public T find(Object id) {
         return getEntityManager().find(entityClass, id);
@@ -60,5 +98,4 @@ public abstract class AbstractFacade<T> {
         javax.persistence.Query q = getEntityManager().createQuery(cq);
         return ((Long) q.getSingleResult()).intValue();
     }
-    
 }
